[Player] Tweak the way we add new genes
When adding new genes, we now use a narrow window around
the previous delta. The objective is to stay in rythm.

When no parents have a meaningful gene to give to the child,
we use a broader window around the previous child delta to explore
and find the new rhythm.

This version of player can now adapt itself to rhythm variation.